Record Equality
The memory tests made it deceptively easy to test that two objects
were the same, as the same object was being pulled out that was
saved. Once I started trying to implement the contract in other
projects, though, I discovered it wasn't so simple when the objects
were being reconstituted.